add new items and groups for recipes

+++ Recipe.cpp
@@ -5,13 +5,13 @@ Recipe::Recipe(string name_t)
 {
     name_of_tool = name_t;
 }
-Recipe::Recipe(string name_t, float p_time, map<string, int> m_consuming) 
+Recipe::Recipe(string name_t, float p_time, map<string, unsigned short int> m_consuming) 
 {
     name_of_tool = name_t;
     production_time = p_time;
     materials_consuming = m_consuming;
 }
-void Recipe::add_material(string name_of_material, int consuming_amount) 
+void Recipe::add_material(string name_of_material, unsigned short int consuming_amount) 
 {
     materials_consuming.insert({name_of_material, consuming_amount});
 }
