+++ tools.hpp
@@ -4,5 +4,5 @@ using namespace std;
 
 void println(string text); 
 void print(string text);
-void print_map_items(map<string, int> collection);
-void print_round(float number_to_print, int precision); 
+void print_map_items(map<string, unsigned short int> collection);
+void print_round(float number_to_print, unsigned short int precision);
